+++ internal/service/route53/service_package_gen.go
@@ -157,6 +157,14 @@ func (p *servicePackage) SDKResources(ctx context.Context) []*inttypes.ServicePa
 			TypeName: "aws_route53_vpc_association_authorization",
 			Name:     "VPC Association Authorization",
 			Region:   inttypes.ResourceRegionDisabled(),
+			Identity: inttypes.GlobalParameterizedIdentity([]inttypes.IdentityAttribute{
+				inttypes.StringIdentityAttribute("zone_id", true),
+				inttypes.StringIdentityAttribute(names.AttrVPCID, true),
+			}),
+			Import: inttypes.SDKv2Import{
+				WrappedImport: true,
+				ImportID:      vpcAssociationAuthorizationImportID{},
+			},
 		},
 		{
 			Factory:  resourceZone,
@@ -196,6 +204,16 @@ func (p *servicePackage) SDKListResources(ctx context.Context) iter.Seq[*inttype
 			},
 				inttypes.WithMutableIdentity(),
 			),
+		},
+		{
+			Factory:  newVPCAssociationAuthorizationResourceAsListResource,
+			TypeName: "aws_route53_vpc_association_authorization",
+			Name:     "VPC Association Authorization",
+			Region:   inttypes.ResourceRegionDisabled(),
+			Identity: inttypes.GlobalParameterizedIdentity([]inttypes.IdentityAttribute{
+				inttypes.StringIdentityAttribute("zone_id", true),
+				inttypes.StringIdentityAttribute(names.AttrVPCID, true),
+			}),
 		},
 		{
 			Factory:  newZoneResourceAsListResource,